refactor: split rust engine into sector/spawn/utils modules
